loop to get new networks from db
